wip: auto-commit from Claude session [skip ci]
