+++ client/src/services/sport/stats.ts
@@ -13,11 +13,50 @@ export interface SportWeekStats {
     streak: number;
 }
 
-function unwrapDisplayValue(value: PrimitiveValue): PrimitiveValue {
+export function unwrapDisplayValue(value: PrimitiveValue): PrimitiveValue {
     if (value.type === PrimitiveValueType.DISPLAY) {
         return value.value.value;
     }
     return value;
+}
+
+/** Build a map of formId -> TIME_ELAPSED question IDs for duration extraction */
+export function buildTimeElapsedFieldsMap(trackables: Trackable[], forms: Form[]): Map<string, string[]> {
+    let formMap = new Map(forms.map(f => [f.id, f]));
+    let fields = new Map<string, string[]>();
+    for (let trackable of trackables) {
+        let form = formMap.get(trackable.formId);
+        if (!form) continue;
+        let qids = form.questions
+            .filter(q => q.dataType === FormQuestionDataType.TIME_ELAPSED)
+            .map(q => q.id);
+        if (qids.length > 0) fields.set(trackable.formId, qids);
+    }
+    return fields;
+}
+
+/** Format a duration in milliseconds to "Xh Ym" string */
+export function formatDurationMs(ms: number): string {
+    let totalMinutes = Math.floor(ms / 60000);
+    let hours = Math.floor(totalMinutes / 60);
+    let minutes = totalMinutes % 60;
+    return `${hours}h ${minutes}m`;
+}
+
+/** Get the total duration in ms for a single journal entry */
+export function getEntryDurationMs(entry: JournalEntry, timeElapsedFields: Map<string, string[]>): number {
+    let fields = timeElapsedFields.get(entry.formId);
+    if (!fields) return 0;
+    let total = 0;
+    for (let fieldId of fields) {
+        let answer = entry.answers[fieldId];
+        if (answer == null) continue;
+        let unwrapped = unwrapDisplayValue(answer);
+        if (unwrapped.type === PrimitiveValueType.NUMBER) {
+            total += unwrapped.value;
+        }
+    }
+    return total;
 }
 
 export class SportStatsService {
@@ -37,41 +76,14 @@ export class SportStatsService {
         weekEnd: Date,
         today: Date
     ): SportWeekStats {
-        let weekStartTs = weekStart.getTime();
-        let weekEndTs = weekEnd.getTime();
+        // Entries are expected to be pre-filtered to the week range by the caller.
+        // We trust that contract instead of re-filtering.
+        let sessions = sportEntries.length;
 
-        let weekEntries = sportEntries.filter(
-            e => e.timestamp >= weekStartTs && e.timestamp <= weekEndTs
-        );
-
-        let sessions = weekEntries.length;
+        let timeElapsedFields = buildTimeElapsedFieldsMap(sportTrackables, forms);
         let totalDurationMs = 0;
-
-        // Build a map of formId -> TIME_ELAPSED question IDs
-        let timeElapsedFields = new Map<string, string[]>();
-        for (let trackable of sportTrackables) {
-            let form = forms.find(f => f.id === trackable.formId);
-            if (!form) continue;
-            let fields = form.questions
-                .filter(q => q.dataType === FormQuestionDataType.TIME_ELAPSED)
-                .map(q => q.id);
-            if (fields.length > 0) {
-                timeElapsedFields.set(trackable.formId, fields);
-            }
-        }
-
-        // Sum durations from all TIME_ELAPSED fields across all week entries
-        for (let entry of weekEntries) {
-            let fields = timeElapsedFields.get(entry.formId);
-            if (!fields) continue;
-            for (let fieldId of fields) {
-                let answer = entry.answers[fieldId];
-                if (answer == null) continue;
-                let unwrapped = unwrapDisplayValue(answer);
-                if (unwrapped.type === PrimitiveValueType.NUMBER) {
-                    totalDurationMs += unwrapped.value;
-                }
-            }
+        for (let entry of sportEntries) {
+            totalDurationMs += getEntryDurationMs(entry, timeElapsedFields);
         }
 
         // Streak is calculated globally (not scoped to week)
